+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Use datetime.timezone.utc instead of relying on pytz for timezone information
 - use toml for config file
+- Automatically install pip in virtual environment. Also install everything from requirements.txt.